new test for AddAspectsAndSetPropertiesProcessor

+++ src/main/java/org/saidone/processors/DeleteNodeProcessor.java
@@ -34,7 +34,7 @@ public class DeleteNodeProcessor extends AbstractNodeProcessor {
     @Override
     public void processNode(String nodeId, Config config) {
         log.debug("deleting node --> {}", nodeId);
-        if (!config.getReadOnly()) {
+        if (config.getReadOnly() != null && !config.getReadOnly()) {
             nodesApi.deleteNode(nodeId, true);
         }
     }
